use-cases/rag-on-gke/rag-e2e/multimodal_emb/requirements.txt
flask
google.cloud.storage
pillow
requests
salesforce-lavis
opencv-python